fix bug for float64; update unit test fn2

+++ torch/_inductor/codecache.py
@@ -2282,12 +2282,12 @@ class CppPythonBindingsCodeCache(CppCodeCache):
             static_assert(std::is_pointer<T>::value, "arg type must be pointer, long, or float");
             return static_cast<T>(_torchinductor_pyobject_tensor_data_ptr(PyTuple_GET_ITEM(args, n)));
         }
-        template <> inline float parse_arg<float>(PyObject* args, size_t n) {
+        template <> inline double parse_arg<double>(PyObject* args, size_t n) {
             auto result = PyFloat_AsDouble(PyTuple_GET_ITEM(args, n));
             if (result == -1.0 && PyErr_Occurred()) {
                 throw std::runtime_error("expected float arg");
             }
-            return static_cast<float>(result);
+            return static_cast<double>(result);
         }
         template <> inline long parse_arg<long>(PyObject* args, size_t n) {
             auto result = PyLong_AsSsize_t(PyTuple_GET_ITEM(args, n));
